requirements.txt
Flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.1.0
chromadb==0.4.22
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
psutil==5.9.6